requirements.txt
# Audio I/O
sounddevice>=0.4.6
scipy>=1.11.0

# Local Whisper (Speech-to-Text) - runs locally, no API key needed
faster-whisper>=1.0.0

# Text-to-Speech
elevenlabs>=1.0.0

# Utils
python-dotenv>=1.0.0
numpy>=1.24.0
